Add textured rendering and material system overhaul
- Added DirectXTex for image loading and as a git submodule
- Implemented Texture and Sampler classes for GPU resource management
- Introduced Material system encapsulating Shader, Texture, and Sampler
- Updated Mesh/Vertex to use UVs; replaced color attributes
- Replaced shaders with HLSL versions supporting texture mapping
- Added sample texture asset and ensured build output copying
- Updated project files for new sources, shaders, and assets
- Improved camera/object controls and updated README with showcase

+++ LostInGalaxy/src/Graphics/ConstantBuffer.hpp
@@ -6,7 +6,7 @@
 template<typename T>
 class ConstantBuffer final : public IBind {
 public:
-	ConstantBuffer( GraphicsDevice* graphicsDevice, uint32_t slot ) :
+	ConstantBuffer( GraphicsDevice* graphicsDevice, uint16_t slot ) :
 		IBind( graphicsDevice ),
 		m_slot( slot )
 	{
@@ -41,13 +41,13 @@ public:
 		m_device->GetContext()->VSSetConstantBuffers( m_slot, 1u, m_buffer.GetAddressOf() );
 	}
 
-	uint32_t GetSlot() const noexcept
+	uint16_t GetSlot() const noexcept
 	{
 		return m_slot;
 	}
 
 private:
 	Microsoft::WRL::ComPtr<ID3D11Buffer> m_buffer;
-	const uint32_t m_slot;
+	const uint16_t m_slot;
 
 };
